pyproject.toml
[tool.poetry]
name = "group-4-image-classification-web-service"
version = "0.1.0"
description = ""
authors = ["dspuliaiev <d.s.puliaiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "^5.1"
Pillow = "^9.4.0"

python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
gevent = "^24.2.1"
tensorflow = "^2.17.0"
tensorflow-io-gcs-filesystem = "0.31.0"
cloud-tpu-client = "^0.10"
cloudinary = "^1.41.0"
django-cloudinary-storage = "^0.3.0"
aiohttp = "^3.10.5"
asyncio = "^3.4.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
